fix: resolve circular dependency in highlighter
